refactor: #149 - 링크 생성 api 함수 파일 이동
