--- src/app/api/link/route.ts
@@ -1,24 +1,0 @@
-import { useServerCookie } from '@/hooks/useServerCookie'
-import { apiServer } from '@/services/apiServices'
-import { NextRequest, NextResponse } from 'next/server'
-
-export async function POST(req: NextRequest) {
-  const { token } = useServerCookie()
-  const { spaceId, url, title, tagName, color } = await req.json()
-  const path = `/spaces/${spaceId}/links`
-  const body = { url, title, tagName, color }
-  const headers = {
-    Authorization: `Bearer ${token}`,
-  }
-
-  try {
-    const response = await apiServer.post(path, body, {}, headers)
-    const data = await response.json()
-    return NextResponse.json(data)
-  } catch (error: any) {
-    return NextResponse.json(
-      { error: error.response.data.message },
-      { status: error.response.status },
-    )
-  }
-}
