+++ src/app/api/space/[spaceId]/links/route.ts
@@ -29,6 +29,27 @@ export async function GET(
   }
 }
 
+export async function POST(req: NextRequest) {
+  const { token } = useServerCookie()
+  const { spaceId, url, title, tagName, color } = await req.json()
+  const path = `/spaces/${spaceId}/links`
+  const body = { url, title, tagName, color }
+  const headers = {
+    Authorization: `Bearer ${token}`,
+  }
+
+  try {
+    const response = await apiServer.post(path, body, {}, headers)
+    const data = await response.json()
+    return NextResponse.json(data)
+  } catch (error: any) {
+    return NextResponse.json(
+      { error: error.response.data.message },
+      { status: error.response.status },
+    )
+  }
+}
+
 export async function DELETE(req: NextRequest) {
   const { token } = useServerCookie()
   const { searchParams } = new URL(req.url)
